scripts/requirements.txt
# Python dependencies for TTS generation
google-genai>=0.2.0
